Adding the hex.py file (still an error)
An error arises when the key is multiple characters long, which is due to my process for xor'ing it taking the ord of a character-long key to raise the input ord to the power of it. I think I misunderstood something regarding how xor'ing works because I'm unsure what else to do with this at the moment. I'm researching it online to try and find the answer. This works fine for operations with single character keys.